Connect apps to layout:   - created 2 more Apps (adapted Geography and Editor to the interface)     - now both apps implement the Application interface     - adapted 2 more routes for Geography and Editor   - refactored existing code a bit     - removed unnecessary data and variables
